split module token rewrite into dedicated microcrate
